moved coloring into baseline utils
